Moved patching to a separate class. Also added check to make sure not to use mod when the game is public lobby
